Move out planting logic
